// src/types/kyc.ts
export interface IUpgradeToLevelOneForm {
  first_name: string;
  last_name: string;
  birthday: string;
  national_code: string;
  day?: string;
  month?: string;
  year?: string;
}
export interface IFaceImageModel {
  face_image: string;
}

export interface IFaceVideoModel {
  face_video: string;
}
export interface ICardNumber {
  card_number: string;
}